+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog]
 
 ## [Unreleased]
 
+- Fix a bug displaying incorrect number of decimal places on monetary number
+  form fields
 - Restrict access to `/admin` by IP
 - Fix student_loan_start_date validation error message